Move input validation to MagicProxyPrinter
