+++ ConsoleApp/Program.cs
@@ -19,7 +19,6 @@ internal class Program
         var serviceProvider = DependencyInjectionConfigurator.Setup();
 
         CoconoaApp.Run(([CoconoaOptions(Description = "Filepath to exported deck from Archidekt")] string? deckFilePath,
-            [CoconoaOptions(Description = "ID of the deck in Archidekt")] int? deckId,
             [CoconoaOptions(Description = "URL link to deck in Archidekt")]string? deckUrl,
             [CoconoaOptions(Description = "Set language for all cards to print")] string? languageCode = null,
             [CoconoaOptions(Description = "Number of copy for each token")] int? tokenCopies = null,
@@ -28,37 +27,10 @@ internal class Program
             [CoconoaOptions(Description = "Filename of the output word file")]string? outputFileName = null,
             [CoconoaOptions(Description = "Flag to store original images in the same folder as output file")] bool storeOriginalImages = false) =>
         {
-            if (deckFilePath is not null)
-            {
-                if (!Path.Exists(deckFilePath))
-                {
-                    ConsoleUtility.WriteErrorMessage("You have to specify correct PATH to your card list exported from Archidekt.");
-                    return;
-                }
-            }
-            else if (deckId is not null)
-            {
-                if (deckId <= 0)
-                {
-                    ConsoleUtility.WriteErrorMessage("You have to specify correct ID of your deck in Archidekt.");
-                    return;
-                }
-            }
-            else if (deckUrl is not null)
-            {
-                var archidektService = serviceProvider.GetService<IArchidektService>()!;
-                if (!archidektService.TryExtractDeckIdFromUrl(deckUrl, out var urlDeckId) || urlDeckId <= 0)
-                {
-                    ConsoleUtility.WriteErrorMessage("You have to specify correct URL to your deck hosted by Archidekt.");
-                    return;
-                }
-                deckId = urlDeckId;
-            }
-            else
+            if (deckUrl is null && deckUrl is null)
             {
                 ConsoleUtility.WriteErrorMessage(@"You have to provide at least one from this list:
                 - path to exported deck
-                - deck id
                 - url to your deck.");
                 return;
             }
@@ -84,7 +56,7 @@ internal class Program
 
             var archidektPrinter = serviceProvider.GetService<IMagicProxyPrinter>()!;
             archidektPrinter.ProgressUpdate += UpdateProgressOnConsole;
-            archidektPrinter.GenerateWord(deckId, 
+            archidektPrinter.GenerateWord(deckUrl, 
                 deckFilePath, 
                 outputPath, 
                 outputFileName,
